pyproject.toml
[tool.poetry]
name = "simplesolutions-task"
version = "0.1.0"
description = "Test task for 'Simple Solutions' company."
authors = ["Alexandr Abramov <https://github.com/Abramov0Alexandr>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^5.0"
djangorestframework = "^3.14.0"


[tool.poetry.group.dev.dependencies]
python-dotenv = "^1.0.0"
django-extensions = "^3.2.3"
psycopg = "^3.1.16"
ipython = "^8.19.0"
stripe = "^7.10.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
